+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 - [GUI] Allow GUI users to delete registry lockfiles (#3829 by: HebaruSan; reviewed: techman83)
+- [GUI] Show unmanaged files in game folder (#3833 by: HebaruSan; reviewed: techman83)
 
 ### Bugfixes